Made changes to Tile and World class
Tile class now has a method to kill the organism and get rows and cols.

World class now uses two arraylists to detect alive and empty neighbours, completed reproduction for a center organism with two neighbours, still need to implement the same algorithm to center organisms with 3 neighbours. This time it needs to randomly select a tile from 3 rather than 2 so its algorithm will be a bit different.