Changed world and tile
World class now makes use of tiles being selected to be able to detect tiles that will have organisms placed to them as a result of reproduction, and does not count them as alive or empty neighbours. It now counts them as selectedTiles and keeps count of it. I have realized a msitake in the 113th line of the code and added what should be done in order to fix it. Will attend to that later. Minor change in the tile class now initializes selected to false in constructor.