Added getColor but not setColor to tile
Adding setColor to Tile makes it so that you have to manually set the color of the tile everytime you add an organism or kill one. Rather than doing that, I have decided to set the color of the organism when a new a organism is added to tile, and set it to transparent when the organism dies.

Also added getColor to Tile class and new comments that concerns the GameManager class for placing the initial organisms.