Some changes after errors in testing
Offsprings are now a property of the class to have a better performance, both in memory and in speed. It was also done this way to be able to get past null pointer exceptions as the offsprings array would need to be initialized every time the nextTurn method runs, which is a pain in the... for performance. More testing coming soon, so probably more changing... This probably still doesn't work as I have encountered a problem in testing that still exists.