Finalizes the World class, a minor but important change to Organism class
World class now randomly selects and places 3 organisms that are neighbour to each other when the world class initializes. It also makes use of the survival chance of each organism to detect if they will survive. Also a temporary isGameOver method is added, which ends the game if a certain amount of rounds is reached (30 in this case) or there are no more organisms left in the grid. Finally, removed a single line from Organism class which could have been bad for the algorithm of the World class.